% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_add_message}
\alias{conv_add_message}
\title{Add a message to current conversation}
\usage{
conv_add_message(x, ...)
}
\description{
Add a message to current conversation
}
\keyword{internal}
