% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_load_and_set}
\alias{conv_load_and_set}
\title{Load a conversation from disk and set as current}
\usage{
conv_load_and_set(x, ...)
}
\description{
Load a conversation from disk and set as current
}
\keyword{internal}
