% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_pending_refresh_data}
\alias{conv_pending_refresh_data}
\title{Get pending refresh data frames}
\usage{
conv_pending_refresh_data(x, ...)
}
\description{
Get pending refresh data frames
}
\keyword{internal}
