% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_pending_refresh_files}
\alias{conv_pending_refresh_files}
\title{Get pending refresh files}
\usage{
conv_pending_refresh_files(x, ...)
}
\description{
Get pending refresh files
}
\keyword{internal}
