% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_set_pending_refresh_data}
\alias{conv_set_pending_refresh_data}
\title{Set pending refresh data frames}
\usage{
conv_set_pending_refresh_data(x, ...)
}
\description{
Set pending refresh data frames
}
\keyword{internal}
