% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_set_pending_refresh_files}
\alias{conv_set_pending_refresh_files}
\title{Set pending refresh files}
\usage{
conv_set_pending_refresh_files(x, ...)
}
\description{
Set pending refresh files
}
\keyword{internal}
