% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_set_pending_refresh_skills}
\alias{conv_set_pending_refresh_skills}
\title{Set pending refresh skills}
\usage{
conv_set_pending_refresh_skills(x, ...)
}
\description{
Set pending refresh skills
}
\keyword{internal}
