% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-conversation.R
\name{conv_set_sent_context_files}
\alias{conv_set_sent_context_files}
\title{Set sent context files}
\usage{
conv_set_sent_context_files(x, ...)
}
\description{
Set sent context files
}
\keyword{internal}
