% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-utils.R
\name{.add_file_to_tree}
\alias{.add_file_to_tree}
\title{Helper to add a single file to the tree}
\usage{
.add_file_to_tree(tree, file)
}
\description{
Helper to add a single file to the tree
}
\keyword{internal}
