% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-ui.R
\name{.build_cassidy_shiny_app}
\alias{.build_cassidy_shiny_app}
\title{Build the Cassidy Shiny app object}
\usage{
.build_cassidy_shiny_app(
  theme,
  context_level,
  context_text,
  assistant_id,
  api_key,
  timeout
)
}
\description{
Build the Cassidy Shiny app object
}
\keyword{internal}
