% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-file.R
\name{.build_file_index_context}
\alias{.build_file_index_context}
\title{Build index-level file context}
\usage{
.build_file_index_context(path, all_lines, total_lines, ext)
}
\description{
Build index-level file context
}
\keyword{internal}
