% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_CHAR_LIMIT_SINGLE}
\alias{.CASSIDY_CHAR_LIMIT_SINGLE}
\title{Conservative single message character limit}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.CASSIDY_CHAR_LIMIT_SINGLE
}
\description{
Conservative single message character limit
}
\keyword{internal}
