% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_CHAR_LIMIT_THREAD}
\alias{.CASSIDY_CHAR_LIMIT_THREAD}
\title{Empirical thread character limit before token failure}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.CASSIDY_CHAR_LIMIT_THREAD
}
\description{
Empirical thread character limit before token failure
}
\keyword{internal}
