% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_LARGE_INPUT_THRESHOLD}
\alias{.CASSIDY_LARGE_INPUT_THRESHOLD}
\title{Large input threshold (100k characters)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.CASSIDY_LARGE_INPUT_THRESHOLD
}
\description{
Large input threshold (100k characters)
}
\keyword{internal}
