% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_TIMEOUT_ERROR_PATTERNS}
\alias{.CASSIDY_TIMEOUT_ERROR_PATTERNS}
\title{Timeout error detection patterns}
\format{
An object of class \code{character} of length 3.
}
\usage{
.CASSIDY_TIMEOUT_ERROR_PATTERNS
}
\description{
Timeout error detection patterns
}
\keyword{internal}
