% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_TOKEN_LIMIT}
\alias{.CASSIDY_TOKEN_LIMIT}
\title{Token limit for CassidyAI API}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.CASSIDY_TOKEN_LIMIT
}
\description{
Token limit for CassidyAI API
}
\keyword{internal}
