% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_VERY_LARGE_INPUT_THRESHOLD}
\alias{.CASSIDY_VERY_LARGE_INPUT_THRESHOLD}
\title{Very large input threshold (250k characters)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.CASSIDY_VERY_LARGE_INPUT_THRESHOLD
}
\description{
Very large input threshold (250k characters)
}
\keyword{internal}
