% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-tokens.R
\docType{data}
\name{.CASSIDY_WARNING_AT}
\alias{.CASSIDY_WARNING_AT}
\title{Warning threshold (fraction of token limit)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.CASSIDY_WARNING_AT
}
\description{
Warning threshold (fraction of token limit)
}
\keyword{internal}
