% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-console.R
\name{.chat_continue_conversation}
\alias{.chat_continue_conversation}
\title{Continue existing conversation}
\usage{
.chat_continue_conversation(message, conv_id, api_key, timeout)
}
\description{
Continue existing conversation
}
\keyword{internal}
