% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-context-handlers.R
\name{.get_env_dataframes}
\alias{.get_env_dataframes}
\title{Get data frames from global environment}
\usage{
.get_env_dataframes(envir = globalenv())
}
\description{
Get data frames from global environment
}
\keyword{internal}
