% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-file.R
\name{.get_language_from_ext}
\alias{.get_language_from_ext}
\title{Get language from file extension}
\usage{
.get_language_from_ext(ext)
}
\description{
Get language from file extension
}
\keyword{internal}
