% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-memory.R
\name{.get_memory_dir}
\alias{.get_memory_dir}
\title{Get memory directory path}
\usage{
.get_memory_dir()
}
\value{
Character. Path to \verb{~/.cassidy/memory/} directory
}
\description{
Returns the path to the user's memory directory. Creates the directory
if it doesn't exist.
}
\keyword{internal}
