% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-context-gather.R
\name{gather_chat_context}
\alias{gather_chat_context}
\title{Gather context for chat app}
\usage{
gather_chat_context(context_level, include_data, include_files)
}
\description{
Gather context for chat app
}
\keyword{internal}
