% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-context-apply.R
\name{setup_apply_context_handler}
\alias{setup_apply_context_handler}
\title{Setup apply context handler}
\usage{
setup_apply_context_handler(
  input,
  session,
  conv_manager,
  assistant_id,
  api_key,
  timeout
)
}
\description{
Setup apply context handler
}
\keyword{internal}
