% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-context-handlers.R
\name{setup_context_data_renderer}
\alias{setup_context_data_renderer}
\title{Setup context data renderer}
\usage{
setup_context_data_renderer(output, input, conv_manager)
}
\description{
Setup context data renderer
}
\keyword{internal}
