% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-context-skills.R
\name{setup_context_skills_renderer}
\alias{setup_context_skills_renderer}
\title{Setup skills context renderer}
\usage{
setup_context_skills_renderer(output, input, conv_manager)
}
\description{
Setup skills context renderer
}
\keyword{internal}
