% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-conversation.R
\name{setup_conversation_list_renderer}
\alias{setup_conversation_list_renderer}
\title{Setup conversation list renderer}
\usage{
setup_conversation_list_renderer(output, conv_manager)
}
\description{
Setup conversation list renderer
}
\keyword{internal}
