% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-context-files.R
\name{setup_file_context_handlers}
\alias{setup_file_context_handlers}
\title{Setup file context handlers}
\usage{
setup_file_context_handlers(
  input,
  session,
  conv_manager,
  assistant_id,
  api_key,
  timeout
)
}
\description{
Setup file context handlers
}
\keyword{internal}
