% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-context-files.R
\name{setup_file_tree_renderer}
\alias{setup_file_tree_renderer}
\title{Setup file tree renderer}
\usage{
setup_file_tree_renderer(output, input, conv_manager)
}
\description{
Setup file tree renderer
}
\keyword{internal}
