% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-context-handlers.R
\name{setup_refresh_context_handler}
\alias{setup_refresh_context_handler}
\title{Setup refresh context handlers}
\usage{
setup_refresh_context_handler(
  input,
  session,
  conv_manager,
  assistant_id,
  api_key,
  timeout
)
}
\description{
Setup refresh context handlers
}
\keyword{internal}
