% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-message.R
\name{setup_send_message_handler}
\alias{setup_send_message_handler}
\title{Setup send message handler}
\usage{
setup_send_message_handler(
  input,
  session,
  conv_manager,
  assistant_id,
  api_key,
  timeout
)
}
\description{
Setup send message handler
}
\keyword{internal}
