% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat-handlers-message.R
\name{setup_message_renderer}
\alias{setup_message_renderer}
\title{Setup message rendering handlers}
\usage{
setup_message_renderer(output, conv_manager)
}
\description{
Setup message rendering handlers
}
\keyword{internal}
